Fixed scroll to error on enter key press
