+++ src/components/formik/Form/ScrollToErrorField/index.js
@@ -2,7 +2,7 @@ import { useEffect, useRef } from "react";
 
 import { useFormikContext } from "formik";
 
-import { getErrorFieldName } from "./utils";
+import { scrollToError } from "./utils";
 
 const ScrollToErrorField = ({ formRef }) => {
   const { submitCount, isValid, errors } = useFormikContext();
@@ -16,18 +16,8 @@ const ScrollToErrorField = ({ formRef }) => {
     if (!formRef.current || isValidatedReference.current || isValid) return;
     isValidatedReference.current = true;
 
-    const fieldErrorName = getErrorFieldName(errors);
-    if (!fieldErrorName) return;
-
-    const errorFormElement = formRef.current.querySelector(
-      `[name="${fieldErrorName}"]`
-    );
-
-    errorFormElement?.scrollIntoView({
-      behavior: "smooth",
-      block: "center",
-    });
-  }, [submitCount, errors]);
+    scrollToError(formRef, errors);
+  }, [submitCount]);
 
   return null;
 };
